+++ backend/cmd/init.go
@@ -7,7 +7,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/AlecAivazis/survey/v2"
+	"go.etcd.io/etcd/client/v3"
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
 	"github.com/sensu/sensu-go/backend"
 	"github.com/sensu/sensu-go/backend/etcd"
@@ -15,20 +15,21 @@ import (
 	etcdstore "github.com/sensu/sensu-go/backend/store/etcd"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"go.etcd.io/etcd/client/v3"
 	"go.etcd.io/etcd/client/pkg/v3/transport"
 	"google.golang.org/grpc"
+	"github.com/AlecAivazis/survey/v2"
 )
 
 const (
 	defaultTimeout = "5s"
 
 	flagIgnoreAlreadyInitialized = "ignore-already-initialized"
-	flagInitAdminUsername = "cluster-admin-username"
-	flagInitAdminPassword = "cluster-admin-password"
-	flagInteractive       = "interactive"
-	flagTimeout           = "timeout"
-	flagWait              = "wait"
+	flagInitAdminUsername        = "cluster-admin-username"
+	flagInitAdminPassword        = "cluster-admin-password"
+	flagInteractive              = "interactive"
+	flagTimeout                  = "timeout"
+	flagWait                     = "wait"
+	flagInitAdminAPIKey          = "cluster-admin-api-key"
 )
 
 var errEtcdEndpointUnreachable = errors.New("etcd endpoint could not be reached")
@@ -41,7 +42,7 @@ type initConfig struct {
 
 func (c *initConfig) Validate() error {
 	if c.SeedConfig.AdminUsername == "" || c.SeedConfig.AdminPassword == "" {
-		return fmt.Errorf("both %s and %s are required to be set", flagInitAdminUsername, flagInitAdminPassword)
+		return fmt.Errorf("both %s and %s are required to be set (or an API key)", flagInitAdminUsername, flagInitAdminPassword)
 	}
 	return nil
 }
@@ -50,6 +51,7 @@ type initOpts struct {
 	AdminUsername             string `survey:"cluster-admin-username"`
 	AdminPassword             string `survey:"cluster-admin-password"`
 	AdminPasswordConfirmation string `survey:"cluster-admin-password-confirmation"`
+	AdminAPIKey 			  string `survey:"cluster-admin-api-key"`
 }
 
 func (i *initOpts) administerQuestionnaire() error {
@@ -74,6 +76,12 @@ func (i *initOpts) administerQuestionnaire() error {
 				Message: "Retype Cluster Admin Password:",
 			},
 			Validate: survey.Required,
+		},
+		{
+			Name: "cluster-admin-api-key",
+			Prompt: &survey.Input{
+				Message: "Cluster Admin API Key:",
+			},
 		},
 	}
 
@@ -141,7 +149,7 @@ func InitCommand() *cobra.Command {
 			}
 
 			timeout := viper.GetDuration(flagTimeout)
-			if timeout < 1 * time.Second {
+			if timeout < 1*time.Second {
 				timeout = timeout * time.Second
 			}
 
@@ -150,6 +158,7 @@ func InitCommand() *cobra.Command {
 				SeedConfig: seeds.Config{
 					AdminUsername: viper.GetString(flagInitAdminUsername),
 					AdminPassword: viper.GetString(flagInitAdminPassword),
+					AdminAPIKey:   viper.GetString(flagInitAdminAPIKey),
 				},
 				Timeout: timeout,
 			}
@@ -166,6 +175,7 @@ func InitCommand() *cobra.Command {
 				}
 				initConfig.SeedConfig.AdminUsername = opts.AdminUsername
 				initConfig.SeedConfig.AdminPassword = opts.AdminPassword
+				initConfig.SeedConfig.AdminAPIKey = opts.AdminAPIKey
 			}
 
 			if err := initConfig.Validate(); err != nil {
@@ -211,6 +221,7 @@ func InitCommand() *cobra.Command {
 	cmd.Flags().Bool(flagInteractive, false, "interactive mode")
 	cmd.Flags().String(flagTimeout, defaultTimeout, "duration to wait before a connection attempt to etcd is considered failed (must be >= 1s)")
 	cmd.Flags().Bool(flagWait, false, "continuously retry to establish a connection to etcd until it is successful")
+	cmd.Flags().String(flagInitAdminAPIKey, "", "cluster admin API key")
 
 	setupErr = handleConfig(cmd, os.Args[1:], false)